lkl: hijack: move dbg.c code into dbg_handler.c
dbg_entrance() is only called via the SIGTSTP signal handler setup in
lkl_register_dbg_handler().
