+++ tools/lkl/lib/Build
@@ -11,7 +11,6 @@ liblkl-$(LKL_HOST_CONFIG_NT) += nt-host.o
 liblkl-y += utils.o
 liblkl-y += virtio_blk.o
 liblkl-y += virtio.o
-liblkl-y += dbg.o
 liblkl-$(LKL_HOST_CONFIG_VIRTIO_NET) += virtio_net.o
 liblkl-$(LKL_HOST_CONFIG_VIRTIO_NET_FD) += virtio_net_fd.o
 liblkl-$(LKL_HOST_CONFIG_VIRTIO_NET_FD) += virtio_net_tap.o
